refactor: move api scraping logic to shared
