+++ apps/api/pyproject.toml
@@ -3,28 +3,17 @@ name = "fastfetchbot-api"
 version = "0.1.0"
 requires-python = ">=3.12,<3.13"
 dependencies = [
-    "fastfetchbot-shared",
+    "fastfetchbot-shared[scrapers]",
     "fastapi>=0.115.12",
     "sentry-sdk[fastapi]>=2.27.0",
     "gunicorn>=23.0.0",
     "uvicorn>=0.34.2",
-    "jinja2>=3.1.6",
     "babel>=2.17.0",
     "beanie>=1.29.0",
-    "jmespath>=1.0.1",
-    "twitter-api-client-v2>=0.1.1",
-    "atproto>=0.0.61",
-    "asyncpraw>=7.8.1",
     "pillow>=10.0.0",
     "pydub>=0.25.1",
     "xhtml2pdf>=0.2.17",
     "aioboto3>=13.4.0",
-    "tenacity>=9.1.2",
-    "markdown>=3.8",
-    "openai>=2.15.0",
-    "html-telegraph-poster-v2>=0.2.5",
-    "firecrawl-py>=4.13.0",
-    "zyte-api>=0.8.1",
     "celery[redis]>=5.4.0",
 ]
