+++ packages/shared/fastfetchbot_shared/services/scrapers/templates/douban_content.jinja2
