+++ packages/shared/fastfetchbot_shared/services/scrapers/templates/douban_short_text.jinja2
