+++ packages/shared/fastfetchbot_shared/services/scrapers/templates/reddit_content.jinja2
