+++ packages/shared/fastfetchbot_shared/services/scrapers/templates/reddit_short_text.jinja2
