+++ packages/shared/fastfetchbot_shared/services/scrapers/templates/video_info.jinja2
