+++ packages/shared/fastfetchbot_shared/services/scrapers/templates/weibo_short_text.jinja2
