+++ packages/shared/fastfetchbot_shared/services/scrapers/templates/xiaohongshu_content.jinja2
