+++ packages/shared/fastfetchbot_shared/services/scrapers/templates/xiaohongshu_short_text.jinja2
