+++ packages/shared/fastfetchbot_shared/services/scrapers/templates/zhihu_short_text.jinja2
